Avoid FP32 intermediate in BF16 modulation ops (~117ms savings on MI300X)
When norm output is BF16 and scale/shift are FP32, the expression
  (norm * (1 + scale_fp32) + shift_fp32)
creates a large [B,S,C] FP32 intermediate tensor (≈200MB for S=32760, C=1536)
then immediately casts it back to BF16.

Fix: cast the small [B,1,C] scale/shift to BF16 first, keeping the large
intermediate in BF16 throughout — 2.4x faster for this computation.

Benchmarked on MI300X at Wan2.1 T2V-1.3B shape [32760, 1536]:
- Current (FP32 intermediate): 0.383 ms
- BF16 arithmetic:             0.160 ms (2.4x faster)
- Savings: ~117 ms per 5-step request (≈300 transformer block calls)

Changes:
- WanTransformerBlock.forward: norm1 modulation uses BF16 scale/shift cast
- WanTransformerBlock_VSA.forward: same
- ScaleResidualLayerNormScaleShift.forward: auto-cast when dtypes differ

+++ python/sglang/multimodal_gen/runtime/models/dits/wanvideo.py
@@ -385,9 +385,10 @@ class WanTransformerBlock(nn.Module):
         # 1. Self-attention
         # Pass hidden_states directly (without .float()) so FP32LayerNorm can
         # use AITER BF16 layernorm on AMD ROCm (~5x faster than FP32 path).
-        # scale_msa/shift_msa are FP32, so type promotion gives FP32 product.
+        # Cast FP32 scale_msa/shift_msa to BF16 before the large [B,S,C]
+        # multiply to avoid a 200MB FP32 intermediate (~2.4x faster).
         norm1 = self.norm1(hidden_states)
-        norm_hidden_states = (norm1 * (1 + scale_msa) + shift_msa).to(orig_dtype)
+        norm_hidden_states = norm1 * (1 + scale_msa).to(orig_dtype) + shift_msa.to(orig_dtype)
         query, _ = self.to_q(norm_hidden_states)
         key, _ = self.to_k(norm_hidden_states)
         value, _ = self.to_v(norm_hidden_states)
@@ -552,10 +553,11 @@ class WanTransformerBlock_VSA(nn.Module):
         # 1. Self-attention
         # Pass hidden_states directly (without .float()) so FP32LayerNorm can
         # use AITER BF16 layernorm on AMD ROCm (~5x faster than FP32 path).
-        # scale_msa/shift_msa are FP32, so type promotion gives FP32 product.
+        # Cast FP32 scale_msa/shift_msa to BF16 before the large [B,S,C]
+        # multiply to avoid a 200MB FP32 intermediate (~2.4x faster).
         norm_hidden_states = (
-            self.norm1(hidden_states) * (1 + scale_msa) + shift_msa
-        ).to(orig_dtype)
+            self.norm1(hidden_states) * (1 + scale_msa).to(orig_dtype) + shift_msa.to(orig_dtype)
+        )
         query, _ = self.to_q(norm_hidden_states)
         key, _ = self.to_k(norm_hidden_states)
         value, _ = self.to_v(norm_hidden_states)
